+++ Assets/_Project/Code/Runtime/Input/InputSystem_Actions.cs
@@ -497,7 +497,7 @@ namespace MrMoonlight.Input
                     ""interactions"": """",
                     ""processors"": """",
                     ""groups"": "";Keyboard&Mouse"",
-                    ""action"": ""Interact"",
+                    ""action"": ""FlashlightToggle"",
                     ""isComposite"": false,
                     ""isPartOfComposite"": false
                 },
@@ -608,17 +608,6 @@ namespace MrMoonlight.Input
                     ""processors"": """",
                     ""groups"": "";Gamepad"",
                     ""action"": ""EquipMelee"",
-                    ""isComposite"": false,
-                    ""isPartOfComposite"": false
-                },
-                {
-                    ""name"": """",
-                    ""id"": ""aaaaaaa2-0009-4002-8002-aaaaaaaa0009"",
-                    ""path"": ""<Keyboard>/l"",
-                    ""interactions"": """",
-                    ""processors"": """",
-                    ""groups"": "";Keyboard&Mouse"",
-                    ""action"": ""FlashlightToggle"",
                     ""isComposite"": false,
                     ""isPartOfComposite"": false
                 },